backend/requirements.txt
asgiref==3.3.0
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
django-filter==2.4.0
djangorestframework==3.12.1
django-cors-headers==3.5.0
pandas==1.1.4
xhtml2pdf
